SOF-263: Organize UI components within core module
