+++ app/src/main/java/com/vci/vectorcamapp/core/presentation/components/form/DatePickerField.kt
@@ -1,4 +1,4 @@
-package com.vci.vectorcamapp.intake.presentation.components
+package com.vci.vectorcamapp.core.presentation.components.form
 
 import android.app.DatePickerDialog
 import android.icu.util.Calendar
@@ -76,7 +76,9 @@ fun DatePickerField(
                         calendar.get(Calendar.YEAR),
                         calendar.get(Calendar.MONTH),
                         calendar.get(Calendar.DAY_OF_MONTH)
-                    ).show()
+                    ).apply {
+                        datePicker.maxDate = System.currentTimeMillis()
+                    }.show()
                 }
         )
