+++ app/src/main/java/com/vci/vectorcamapp/landing/presentation/components/LandingActionTile.kt
@@ -18,7 +18,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.res.painterResource
 import com.vci.vectorcamapp.R
-import com.vci.vectorcamapp.core.presentation.components.ui.ActionTile
+import com.vci.vectorcamapp.core.presentation.components.tile.ActionTile
 import com.vci.vectorcamapp.ui.extensions.colors
 import com.vci.vectorcamapp.ui.extensions.dimensions
